% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{get_gargoyle_logs}
\alias{get_gargoyle_logs}
\alias{clear_gargoyle_logs}
\title{Handle logs}
\usage{
get_gargoyle_logs()

clear_gargoyle_logs()
}
\value{
a data.frame
}
\description{
Get / Clear the logs of all the time the `trigger()` functions are launched.
}
\examples{
if (interactive()){
  get_gargoyle_logs()
  clear_gargoyle_logs()
}
}
