% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{new_data}
\alias{new_data}
\alias{new_listeners}
\title{Create data / listeners}
\usage{
new_data(...)

new_listeners(...)
}
\arguments{
\item{...}{for data, a named list with name / value. For
listeners, the names of the listeners.}
}
\value{
An environment.
}
\description{
Create data / listeners
}
