% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{init}
\alias{init}
\alias{trigger}
\alias{watch}
\title{Initiate, triger, event}
\usage{
init(..., session = getDefaultReactiveDomain())

trigger(..., session = getDefaultReactiveDomain())

watch(name, session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The shiny session object}

\item{name, ...}{The name(s) of the events}
}
\description{
Initiate, triger, event
}
