% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{on}
\alias{on}
\title{React on an event}
\usage{
on(name, expr, session = getDefaultReactiveDomain())
}
\arguments{
\item{name}{the name of the event to react to}

\item{expr}{the expression to run when the event
is triggered.}

\item{session}{The shiny session object}
}
\description{
React on an event
}
